+++ src/main/java/my/app/dao/UserDAO.java
@@ -7,12 +7,12 @@ import java.util.List;
 public interface UserDAO {
     List<User> listAllPeople();
 
-    User getUserById(int id);
+    User getUserById(Long id);
 
     void save(User user);
 
-    void update(int id, User updatedUser);
+    void update(Long id, User updatedUser);
 
-    void delete(int id);
+    void delete(Long id);
 }
